TI: firmware: ti_sci: Partial-IO support
Add support for Partial-IO poweroff. In Partial-IO a few modules can
generate a system wakeup.

The modules that can be wakeup sources are defined by a DT property. If
one of these has wakeup enabled, instead of a normal poweroff we are
entering the Partial IO mode to be able to wakeup on activity.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
(cherry picked from commit 241d614821ca8e89f972dbc95047dcabc7f82130)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>