TI: HACK: arm64: dts: ti: Configure am62a pinctrl-single for wake-up interrupts
Configure am62a pinctrl-single for wake-up interrupts.

Based on the patch for am625 with changes to the length of the register
set and the interrupt ID. The length has been updated to match
end address - start address for Main PADCFG registers. The interrupt ID
has been updated to next available line than what was used on AM625.

Marked as HACK because physically none of the wakeup source events are
connected to GIC. Eventually we may want to set up a chained interrupt
handler in the ti-msgmgr for pinctrl-single to use.

Signed-off-by: Vibhore Vardhan <vibhore@ti.com>
(cherry picked from commit 4bba9a21d056a81562268382f23629d85ee96ba2)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>