UPSTREAM: arm64: dts: ti: k3-am625-phyboard-lyra-rdk: Add HDMI support
The DSS outputs DPI signals via its second video port (VP2). The DPI
output from DSS is 24 bits (RGB888) and is forwarded to an HDMI
transmitter (SIL9022) on the board.

Add pinmux for DSS DPI output and HDMI Interrupt.

Add DT nodes for SIL9022 HDMI transmitter (TX), and the HDMI connector
on the phyBOARD-Lyra.

Additionally, connect the output of DSS (VP2) with input of the HDMI TX,
and the output of HDMI TX to the input of the HDMI connector.

Change-Id: Ib5af68c29c701416c3faa2d96c6b78e090a6d188
Signed-off-by: Garrett Giordano <ggiordano@phytec.com>
Reviewed-by: Wadim Egorov <w.egorov@phytec.de>
Link: https://lore.kernel.org/r/20231204222811.2344460-2-ggiordano@phytec.com
Signed-off-by: Nishanth Menon <nm@ti.com>
(cherry picked from commit bac441710306a84c52f0f9a561aa9839b91caa14)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>