UPSTREAM: usb: typec: tipd: fix use of device-specific init function
The current implementation supports device-pecific callbacks for the
init function with a function pointer. The patch that introduced this
feature did not update one call to the tps25750 init function to turn it
into a call with the new pointer in the resume function.

Fixes: d49f90822015 ("usb: typec: tipd: add init and reset functions to tipd_data")
Change-Id: I35353b5e5f11209243f60a1f7760771ddacbca19
Signed-off-by: Javier Carrasco <javier.carrasco.cruz@gmail.com>
Reviewed-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Suggested-by: Roger Quadros <rogerq@kernel.org>
Link: https://lore.kernel.org/r/20240104-dev_spec_init-v1-1-1a57e7fd8cc8@gmail.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 933bb7b878ddd0f8c094db45551a7daddf806e00)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>