UPSTREAM: arm64: dts: ti: k3-j7200: Remove PCIe endpoint node
This node is an example node for the PCIe controller in "endpoint" mode.
By default the controller is in "root complex" mode and there is already a
DT node for the same.

Examples should go in the bindings or other documentation.

Remove this node.

Change-Id: If786ea7ff059974edf369cd44c1c553b1550117b
Signed-off-by: Andrew Davis <afd@ti.com>
Link: https://lore.kernel.org/r/20240124183659.149119-2-afd@ti.com
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
(cherry picked from commit 0b16abe711bd5136f2319c4f548fb20588540266)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>