UPSTREAM: arm64: dts: ti: k3-am64: Remove PCIe endpoint node
This node is an example node for the PCIe controller in "endpoint" mode.
By default the controller is in "root complex" mode and there is already a
DT node for the same.

Examples should go in the bindings or other documentation.

Remove this node.

Change-Id: I13211dc2e0f6a5eb730b431c62f92125db19ab15
Signed-off-by: Andrew Davis <afd@ti.com>
Link: https://lore.kernel.org/r/20240124183659.149119-4-afd@ti.com
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
(cherry picked from commit 6cce60550763564360beadf57ecf2e4d45b585e1)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>