BACKPORT: arm64: dts: ti: k3-j7200: Add support for multiple CAN instances
CAN instances 0 and 1 in the mcu domain are brought on the common
processor board through headers J30 and J31 respectively. Thus, add
their respective transceivers 1 and 2 dt nodes to add support for
these CAN instances.

CAN instance 3 in the main domain is brought on the common
processor board through header J27. The CAN High and Low lines
from the SoC are routed through a mux on the SoM. The select lines need
to be set for the CAN signals to get connected to the transceiver 3 on
the common processor board. Therefore, add transceiver dt nodes to add
support for this CAN instance.

Signed-off-by: Bhavya Kapoor <b-kapoor@ti.com>
Link: https://lore.kernel.org/r/20240130102044.120483-4-b-kapoor@ti.com
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
(cherry picked from commit 6b80695f93b2253dbf669323154eb06b162f55ef)
Conflicts:
   arch/arm64/boot/dts/ti/k3-j7200-common-proc-board.dts
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>