TI: arm64: dts: ti: k3-j722s: Add support for PCIe0
The PCIe0 instance of PCIe on TI's J722S SoC is a Gen3 single lane PCIe
controller.

Add the device-tree nodes for it and enable it in Root Complex mode of
operation using Lane 0 of the Serdes1 instance of Serdes.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
(cherry picked from commit 2b466a25764b90577292416556566d1971a489ad)
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>