UPSTREAM: can: m_can: Introduce a tx_fifo_in_flight counter
Keep track of the number of transmits in flight.

This patch prepares the driver to control the network interface queue
based on this counter. By itself this counter be
implemented with an atomic, but as we need to do other things in the
critical sections later I am using a spinlock instead.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Link: https://lore.kernel.org/all/20240207093220.2681425-12-msp@baylibre.com
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>
(cherry picked from commit 1fa80e23c15051edc1c594270517de3517ded798)
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>