TI: ANDROID: arm64: dts: ti: k3-am62p5-sk: add 512MiB memory carveout region for wave5
On AM62Px, the video encoder/decoder (wave5) has a hardware limitation:
For each memory access, the upper 16 bits of the 48 bit address *must* be
identical.

In order to enforce this addressing limitation, we can use a different dmabuf
heap, named carveout[1].

Add a chunk of 512MiB of carveout heap for dedicated video memory.

Notes:
- 512MiB has been picked experimentally, by measuring a decode of
  H264 6k 30FPS video, which used 478MiB of memory.
- the carveout region starts just after OP-TEE, and before our CMA region.

[1] https://git.ti.com/cgit/ti-linux-kernel/ti-linux-kernel/tree/drivers/dma-buf/heaps/carveout-heap.c?h=ti-linux-6.1.y
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>