TI: ANDROID: media: chips-media: wave5: support 48bit addr via carveout dma_heap
The wave5 chip has a hardware limitation: when using 48bit memory addressing,
the upper 16 bits of each memory access must be static.

These 16 upper bits are written into a register value that cannot
be rewritten to after device initialization finishes.

Since we cannot control in which physical memory region the memory
allocations are happening, we should switch to a different dma-buf heap.
The carveout[1] heap allows to specify a phys_addr_t address and a size,
and will guarantee that all allocs will be in these boundaries.

Add support for the carveout dma-heap. This requires device-tree changes:

Add a reserved-memory node:

   carveout_video: video {
           compatible = "dma-heap-carveout";
           reg = <0x00 0xa0000000 0x00 0x20000000>;
           no-map;
   };

Provide the name to the wave5 driver:

  &vpu {
        ti,dma-heap-name = "carveout_video";
  };

[1] https://git.ti.com/cgit/ti-linux-kernel/ti-linux-kernel/tree/drivers/dma-buf/heaps/carveout-heap.c?h=ti-linux-6.1.y
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>