TI: dma-buf: heaps: carveout-heap: add helper to get phys_addr_t
The carveout heap was made to ensure that all allocations are done in a well
known physical address range of the DDR.

This is useful for devices that have special address requirements.

The wave5 driver is such device, since the upper 16 bits of its 48 bit memory
addresses must be identical for all allocations.

Add a helper function to expose the underlying phys_addr_t so that drivers such
as the wave5 can retrieve that information.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>