TI: remoteproc: core: Add DMA-BUF attachment interface to cdev
Like many devices capable of DMA, a remote processor may need to work
with memory buffers produced by other devices. The standard method for
this is the DMA-BUF framework. The RemoteProc framework already exposes
a character device interface (cdev). Extend the interface provided by
this cdev to allow importing DMA-BUF handles.

The DMA-BUF is attached and mapped to the remote processor using a new
ioctl: DMA_BUF_ATTACH. This handles any bus translation and/or coherency
management.

Communicating this DMA-BUF to the remote processor is still a work in
progress. One option is to treat it how carveouts are currently. The
buffer address and size would be set as part of the resource table on
boot/attach. The issue would be that all buffers would need to be
attached before boot/attach, but we would like to have these dynamically
attached/detached at runtime as needed.

Another gap is how pointers into these buffers should be sent by users to
the remote processors. We do not have a standard in place for sending
memory addresses over rpmsg. So we have no universal way to do any
translation in-flight from userspace pointers into device address
pointers. This means for now userspace will have to send raw device
addresses in the messages it sends. To facilitate this we return a device
address base pointer to userspace when attaching a DMA-BUF. This is only
a temporary solution as userspace should not be given physical addresses
for various reasons.

Change-Id: I4db744857166f9d7c68186ae461ab9566dfd50c9
Signed-off-by: Andrew Davis <afd@ti.com>
(cherry picked from commit cde91a89becee656867b8c8a86e627241dc91449)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>