TI: spi: cadence-quadspi: fix error check when finding rxhigh
During PHY calibration the return code of cqspi_find_rx_high() is not
assigned to 'ret'. So the check for failure will take the previous value
of 'ret', which will always be 0 since it was already checked by the
previous step. This means that the calibration procedure will always
think that rxhigh is found even when it is not. Fix this by assigning
the return value of cqspi_find_rx_high() to 'ret'.

Fixes: be58a6d2184f ("spi: cadence-qspi: Tune PHY to allow running at higher frequencies")
Reported-by: Brian Paiva <bpaiva@blackberry.com>
Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Apurva Nandan <a-nandan@ti.com>
(cherry picked from commit 9ca1870d256cd80bc78d4ea05a0f3da5d1a6db65)
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>