UPSTREAM: mailbox: omap: Move omap_mbox_irq_t into driver
This is only used internal to the driver, move it out of the
public header and into the driver file. While we are here,
this is not used as a bitwise, so drop that and make it a
simple enum type.

Change-Id: Iac76db136e59c3e59485ac3ccd964b072d6e8dd9
Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Jassi Brar <jassisinghbrar@gmail.com>
(cherry picked from commit 6979e8be50af143a373bf5905a176434a5880ca4)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>