UPSTREAM: mailbox: omap: Remove device class
The driver currently creates a new device class "mbox". Then for each
mailbox adds a device to that class. This class provides no file
operations provided for any userspace users of this device class.
It may have been extended to be functional in our vendor tree at
some point, but that is not the case anymore, nor does it matter
for the upstream tree.

Remove this device class and related functions and variables.
This also allows us to switch to module_platform_driver() as
there is nothing left to do in module_init().

Change-Id: Ib3ede24f44184118dece638efe6ff8018552149c
Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Jassi Brar <jassisinghbrar@gmail.com>
(cherry picked from commit 982b1451517df4e8aa8de5042fcd026970a34094)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>