UPSTREAM: mailbox: omap: Remove mbox_chan_to_omap_mbox()
This function only checks if mbox_chan *chan is not NULL, but that cannot
be the case and if it was returning NULL which is not later checked
doesn't save us from this. The second check for chan->con_priv is
completely redundant as if it was NULL we would return NULL just the
same. Simply dereference con_priv directly and remove this function.

Change-Id: I7df0cfd116644dd5e23ed951cefb15278666e520
Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Jassi Brar <jassisinghbrar@gmail.com>
(cherry picked from commit 5aa00b68eadee64e1a6e95325f364511f37631d8)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>