TI: rpmsg: char: Update local endpt address for virtio-rpmsg backend
The rpmsg char driver creates a local end-point when the actual
endpt device is opened. The virtio-rpmsg backend can dynamically allocate
the local end-point address if the endpt creation is done using the
address RPMSG_ADDR_ANY. This is not reflected in the sysfs src file, so
update the stored address with the allocated address in such a case. This
allows the userspace to be able to retrieve the local end-point
address through sysfs.

Change-Id: Ie01fccde4fd5882a0797e9695f39f8c077c497f6
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
(cherry picked from commit 6e3d3b577e2db60c865b7f8af518c789317efcd1)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>