TI: thermal: k3_j72xx_bandgap: implement suspend/resume support
This add suspend-to-ram support.

The derived_table is kept-as is, so the resume is only about
pm_runtime_* calls and restoring the same registers as the probe.

Extract the hardware initialization procedure to a function called at
both probe-time & resume-time.

The probe-time loop is split in two to ensure doing the hardware
initialization before registering thermal zones. That ensures our
callbacks cannot be called while in bad state.

The 100ms delay in the hardware initialization sequence was removed.
It was initially added to be sure the thresholds are programmed before
enabling the interrupt, but in fact it's not needed (tested on J7200
platform).

Change-Id: I3f520eeeda623ab0e24c87e80f91b30c05f4b930
Signed-off-by: ThÃ©o Lebrun <theo.lebrun@bootlin.com>
Acked-by: Keerthy <j-keerthy@ti.com>
Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>
(cherry picked from commit d0008a171895767c99f0d7ef921909e45ca6636c)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>