TI: input: edt-ft5x06: Correct prefix length in snprintf
[commit e4631fd5c3c3c3710636dfdf9657bdb188bec51d rpi-linux]

snprintf takes the length of the array that we can print into,
and has to fit the NULL terminator in there too.
Printing the prefix is generally "12-3456 " which is 8 desired
characters (the length of EDT_NAME_PREFIX_LEN) and the NULL.
The space is therefore being truncated to fit the NULL in.

Increase the length snprintf is allowed to use.

Patch Link:
https://github.com/raspberrypi/linux/commit/e4631fd5c3c3c3710636dfdf9657bdb188bec51d

Change-Id: Id2863ce772e0782ab96a45a8948f4e52cfc26cfb
Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
(cherry picked from commit fa68c2d4748a48bead9344f7cccc119d7caa068e)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>