TI: drm/bridge: cdns-dsi: Wait for Clk and Data Lanes to be ready
Once the DSI Link and DSI Phy are initialized, the code needs to wait
for Clk and Data Lanes to be ready, before continuing configuration.
This is in accordance with the DSI Start-up procedure, found in the
Technical Reference Manual of Texas Instrument's J721E SoC[0] which
houses this DSI TX controller.

If the previous bridge (or crtc/encoder) are configured pre-maturely,
the input signal FIFO gets corrupt. This introduces a color-shift on the
display.

Allow the driver to wait for the clk and data lanes to get ready during
DSI enable.

Fixes: e19233955d9e ("drm/bridge: Add Cadence DSI driver")
Change-Id: I94282aec5ab0ad9758db2ed79c52a84debf97866
Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
(cherry picked from commit de0415e9312c22d906a22715a1b62f9c97c17431)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>