TI: drm/panel: simple: Add VESA timings for Raspberry Pi 7" DSI panel
[Commit 8648a8c9442b45e85b73423be1b469d10234b4e2 rpi-linux]
[Commit bde1b4c52cc5a79a2108076b8706e23d5844afa9 beagle-linux]

The Raspberry Pi 7" 800x480 panel uses a Toshiba TC358762 DSI
to DPI bridge chip, so there is a requirement for the timings
to be specified for the end panel. Add such a definition.

This patch is a squash of the following 2 commits.

Patch Links
https://github.com/raspberrypi/linux/commit/8648a8c9442b45e85b73423be1b469d10234b4e2
https://github.com/beagleboard/linux/commit/bde1b4c52cc5a79a2108076b8706e23d5844afa9

Change-Id: Ie4a3f114cc3c9beee3bf1d935dc21ddf2b925fb3
Signed-off-by: Rahul T R <r-ravikumar@ti.com>
Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
(cherry picked from commit 02e95ec004ef0a1d5021dcd28be35800c547614f)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>