TI: dt-bindings: serial: 8250_omap: Add wakeup pinctrl state
Pins associated with the 8250 omap unit can be the source of a wakeup in
deep sleep states. To be able to wakeup, these pins have to be
configured in a special way. To support this configuration add the
default and wakeup pinctrl states.

Change-Id: I6e58a2a5e45e8c53eabe3fb6f042ac975b54d009
Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
(cherry picked from commit 426ed3f5b3b21fbfee391953c3e9cf3d9dc4831f)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>