TI: arm64: dts: ti: Add overlay for RPi header for AM625-SK and AM62 LP-SK
AM625-SK and AM62 LP-SK have GPIO and ePWM signals routed to the
40 pin header (J3). This overlay will set the appropriate pinmux and
enable GPIO and PWM on those pins.

PWM0 A and B are routed to pin 12 and 40 respectively.
PWM1 A and B are routed to pin 36 and 33 respectively.

The gpio-hog p25 is needed because some pins are behind SoC UART1
FET SWITCH which is default HIGH and needs to be unset to route PWM0
A and B signals to the user expansion header.

The gpio-hog p05 and p06 are need to enable 3.3V and 5V rails on the
user expansion header.

Change-Id: Ic7571fb431c181a006e3c17b9188fd2f5e424ce3
Signed-off-by: Judith Mendez <jm@ti.com>
Reviewed-by: Bryan Brattlof <bb@ti.com>
(cherry picked from commit bb08285811ba071f21d4faec7e41777c9c1417b5)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>