TI: arm64: dts: ti: Add overlay for RPi header for AM62A-SK
AM62A-SK has GPIO and ePWM signals routed to the 40 pin header (J3).
This overlay will set the appropriate pinmux and enable GPIO and
PWM on those pins.

PWM0 A and B are routed to pin 35 and 12 respectively.
PWM1 A and B are routed to pin 36 and 33 respectively.

The gpio-hog p25 is needed because some pins are behind SoC UART1
FET SWITCH which is default HIGH and needs to be unset to route PWM0
A and B signals to the user expansion header.

The gpio-hog p05 and p06 are need to enable 3.3V and 5V rails on the
user expansion header.

Change-Id: I063f316bdb1bdc4aacbb6de70e6756cc03afbf62
Signed-off-by: Judith Mendez <jm@ti.com>
Reviewed-by: Bryan Brattlof <bb@ti.com>
(cherry picked from commit 9efb8f0b337689bfb9194829cfbe32a89d3c33f3)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>