TI: arm64: dts: ti: Add overlay for RPi header for AM62P-SK
AM62P-SK has GPIO and ePWM signals routed to the 40 pin header (J4).
This overlay will set the appropriate pinmux and enable GPIO and
PWM on those pins.

PWM0 A and B are routed to pin 35 and 12 respectively.
PWM1 A and B are routed to pin 36 and 33 respectively.

The gpio-hog p02 is needed because some pins are behind SoC UART1
FET SWITCH which is default HIGH and needs to be unset to route PWM0
A and B signals to the user expansion header.

The gpio-hog p05 is need to enable the 3.3V rail on user expansion
connector. It is enabled by setting EXP_PS_3V3_EN GPIO on the
io-expander at address 0x22.

The gpio-hog p01 is need to enable the 5V rail on user expansion
connector. It is enabled by setting EXP_PS_5V0_EN GPIO on the
io-expander at address 0x23.

Since dss support is not in 6.6 TI Linux kernel, comment out disabling
of dss node for now, this will be fixed once dss support is added.

Change-Id: I11806a7a83e2bd8057eb062a09c0ff1e4b7ad41a
Signed-off-by: Judith Mendez <jm@ti.com>
Reviewed-by: Bryan Brattlof <bb@ti.com>
(cherry picked from commit a45124fdcd2e9624d6bc34abef97c55348ab80f8)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>