TI: drm/tidss: Add dispc_is_idle()
Add a helper function, dispc_is_idle(), which returns whether the DSS is
idle (i.e. is any video port enabled).

For now we add a call to it in the suspend and resume callbacks, and
print a warning if in either place the DSS is not idle.

Change-Id: I3c4717155859bf1aeba364f23facb84ad3930a5a
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
(cherry picked from commit 8e76a266feee641ade27877f476a0c6266668aa3)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>