TI: HACK: drm/panel-edp: make support for DP AUX backlight conditional
DP AUX backlight support was added to create aux backlight if there is
no backlight property in the panel node in the device-tree but panel has
access to aux channel. To do so dpcd reads are performed to get data from
sink(monitor). And if sink is not connected to DP-1, we get timeout errors
due to which  panel and other dependent components are not probed.
So if we have DP-0 connected to display, due to this timeout, tidss is not
probed and display does not work.
Therefore making this backlight support conditional for now till we have
a proper fix. This should not cause any issue since we do not actually
need backlight.

Change-Id: I3b72a7783440e18a0b4306ab4cd49ff1e14166e7
Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
(cherry picked from commit 7e0771332633f2a98a9d826b7c8d4ea59df056a6)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>