TI: arm64: dts: ti: k3-j784s4-evm: Add support for multiple CAN instances
CAN instances 0 and 1 in the mcu domain and 16 in the main domain are
brought on the evm through headers J42, J43 and J46 respectively. Thus,
add their respective transceiver's 0, 1 and 2 dt nodes to add support
for these CAN instances.

CAN instance 4 in the main domain is brought on the evm through header
J45. The CAN High and Low lines from the SoC are routed through a mux
on the evm. The select lines need to be set for the CAN signals to
reach to its transceiver on the evm. Therefore, add transceiver 3
dt node to add support for this CAN instance.

Change-Id: I58c7b1c9e1db85cd438e553303e65d3342c208f7
Signed-off-by: Bhavya Kapoor <b-kapoor@ti.com>
(cherry picked from commit 50f59606626f94ef1fb1772969b613153ae2f361)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>