TI: arm64: boot: dts: ti: k3-*: Add memory node to bootloader stage
Add the bootph-all property to the memory node so that it can be
accessed by FDT functions at bootloader stage.

The bootloader requires the memory node to be able to initialize and set
the size of the DRAM banks. For this purpose, make sure all memory nodes
are present and standardized, and modify them if not.

Change-Id: I16257c7d8678006e7ed66df23872b32e0494d00d
Signed-off-by: Neha Malcom Francis <n-francis@ti.com>
(cherry picked from commit 26e670c5d29f40931827a9a5b76b834687a389a1)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>