TI: media: ti: j721e-csi2rx: add multistream support
Each CSI2 stream can be multiplexed into 4 independent streams, each
identified by its virtual channel number. To capture this multiplexed
stream, the application needs to tell the driver how it wants to route
the data. It needs to specify which context should process which stream.
This is done via the new routing APIs.

Add ioctls to accept routing information from the application and save
that in the driver. This can be used when starting streaming on a
context to determine which route and consequently which virtual channel
it should process.

Use the new enable_streams() APIs to start/stop individual streams on
the connected subdev when capture is initiated on the corresponding v4l2
device.

Co-developed-by: Pratyush Yadav <p.yadav@ti.com>
Change-Id: I731cc54eb6fccf4083a4cb0974d9f78cef37a39c
Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Jai Luthra <j-luthra@ti.com>
(cherry picked from commit 384b9fd6edc500a690811ab5c743ae07fffa7766)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>