TI: HACK: arm64: dts: ti: Configure am62p pinctrl-single for wake-up interrupts
Configure am62p pinctrl-single for wake-up interrupts.

Based on the patch for am625. No changes to the length of the register
set and the interrupt ID were needed for am62p as they match those on
am625.

Marked as HACK because physically none of the wakeup source events are
connected to GIC. Eventually we may want to set up a chained interrupt
handler in the ti-msgmgr for pinctrl-single to use.

Change-Id: I3c0333b0dd65fde4a1eb47a3818a06a3f329237b
Signed-off-by: Vibhore Vardhan <vibhore@ti.com>
Signed-off-by: Dhruva Gole <d-gole@ti.com>
(cherry picked from commit 21354a5e32c40bc97c8e602cb2174dc9ede672d3)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>