TI: counter/ti-eqep: add support for unit timer
This adds support to the TI eQEP counter driver for the Unit Timer.
The Unit Timer is a device-level extension that provides a timer to be
used for speed calculations. The sysfs interface for the Unit Timer is
new and will be documented in a later commit. It contains a R/W time
attribute for the current time, a R/W period attribute for the timeout
period and a R/W enable attribute to start/stop the timer. It also
implements a timeout event on the chrdev interface that is triggered
each time the period timeout is reached.

Link: https://www.spinics.net/lists/linux-iio/msg64817.html

Change-Id: I7cb8e3219038aff5603fe9e39f46fc6b5693069f
Signed-off-by: David Lechner <david@lechnology.com>
[Judith: Only clear addressed interrupts. Change clock to NULL.
Fail probe if get clock rate fails]
Signed-off-by: Judith Mendez <jm@ti.com>
(cherry picked from commit f32b3c9e064ab71e25995b4b40ef144f14e926a4)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>