TI: docs: networking: ti: add driver doc for CPSW Proxy Client
The CPSW Proxy Client driver interfaces with Ethernet Switch Firmware on
a remote core to enable Ethernet functionality for applications running
on Linux. The Ethernet Switch Firmware (EthFw) is in control of the CPSW
Ethernet Switch on the SoC and acts as the Server, offering services to
Clients running on various cores.

Change-Id: Ie915bf4108457bcf490e930733e962fe73e4891d
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
(cherry picked from commit 8509c70d4037b8e87a8077fffcfe025476159be9)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>