TI: net: ethernet: ti: cpsw-proxy-client: enable message exchange with EthFw
Add a new function which sends RPMsg requests to EthFw and shares the
response for the request. The RPMsg callback function copies the response
it receives from EthFw to the driver's private member, thereby allowing
the response to be shared with the newly added function which sends the
request.

Change-Id: Ic46be5498362f05192c7b555cc3164fc58fe1cf9
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
(cherry picked from commit 152152159ac267da221c373a2cb4109b2fca1035)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>