TI: net: ethernet: ti: cpsw-proxy-client: add helpers to (de)register IPv4
Add functions "register_ipv4()" and "deregister_ipv4()" to register and
deregister IPv4 Address of the network interface corresponding to the
Virtual Switch Port with EthFw. Registering the IPv4 Address with EthFw
is necessary in the case of the Virtual Switch Port. This is because all
Broadcast packets received on any of the Switch Ports are consumed by
EthFw. This includes the ARP request for the IPv4 Address of the network
interface corresponding to the Virtual Switch Port as well. Thus,
registering the IPv4 Address with EthFw results in EthFw responding to
the ARP request thereby enabling subsequent Unicast communication with
the network interface corresponding to the Virtual Switch Port.

Add a notifier block to register/deregister the IPv4 address with EthFw
corresponding to interface state changes as well as IPv4 Address
changes.

Change-Id: I6b8bfa72a5d89c790171a263170fc2e74c8ae66e
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
(cherry picked from commit 15c1ba4a5017d845c52c020f7feddae9b6d6946c)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>