TI: media: vxe-vxd: decoder: Error handling of fatal condition
Handling of the fatal error response from the IP, setting
the state of frame to discard and resetting the IP. The error
recovery function is executed in the error simulation condition,
when emergency flag is not set, only in a non fatal condition.
When the emergency flag is set, the IP reset is handled in the
schedule_work() api of Interrupt handling.

Change-Id: I691709163297f36f689dd9a900767b119e51a18c
Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Brandon Brnich <b-brnich@ti.com>
(cherry picked from commit e0daf9e95f3b4787a15a7f5a1222d84db2dd97f7)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>