TI: media: platform: img: vxe-vxd: fix a array out of bounds bug
This changed sometime after kernel 4.19.  One reserved field was
removed, so writing to the v4l2_ext_controls reserved field for
array index 0 and 1 actually is an array out of bounds write.

I am not entirely sure why we are writing to a reserved field in
the first place.  But I'll fix this bug for now, and revisit that
question at a later date.

Change-Id: I0dd47abf09a6abd97696c5d793e62b7316aaac62
Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Brandon Brnich <b-brnich@ti.com>
(cherry picked from commit c558f85808e8fb06e995c11119340514bbe41112)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>