TI: media: platform: img: vxe-vxd: fix size of firmware comms structure
Originally the driver was built as if MVC was disabled in the firmware.

MVC here refers to Multiview Video Coding as defined in the H.264 spec.

However MVC is actually enabled so we need the firmware communication
structures to be bigger to contain the MVC data.  It wasn't causing any
obvious problems previously, but it makes sense to make sure the
structure sizes all match between the host and the firmware so that all
memory is initialized and mapped correctly.

Change-Id: I4b3e4c4dc33c154aaf373e700ecb7f0e40636710
Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Brandon Brnich <b-brnich@ti.com>
(cherry picked from commit 6fb589fe3cf7f66fab4ae4b1be82ef26b50f48db)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>