TI: net: ethernet: ti: icssg_prueth: introduce and use prueth_swdata struct for SWDATA
We have different cases for SWDATA (skb, page, cmd, etc)
so it is better to have a dedicated data structure for that.
We can embed the type field inside the struct and use it
to interpret the data in completion handlers.

Increase SWDATA size to 48 so we have some room to add
more data if required.

Change-Id: Iad00584f24805c7eb9a77f220469f84eeb82b005
Signed-off-by: Roger Quadros <rogerq@kernel.org>
Signed-off-by: MD Danish Anwar <danishanwar@ti.com>
(cherry picked from commit b4d9c256c6b7cf9160ac3e34279fda1564427751)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>