TI: HACK: mtd: spi-nor: core: run calibration when initialization is done
Once the flash is initialized tell the controller it can run
calibration procedures if needed. This can be useful when calibration is
needed to run at higher clock speeds.

This is marked as a hack because it is not the cleanest way of telling
the controller where the pattern is stored. The discussion is going on
upstream on what the best way to do it is. Once a conclusion is reached,
this can be replaced with the proper solution. Anyway, it is better than
how it was done in 5.4 because it works for both device tree and command
line partitions.

Change-Id: Id8765ec2405b6cd67c73f1b8ec6f102fc1c785f9
Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
(cherry picked from commit 54558c826528471b754a554429351ec41927f22a)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>