TI: arm64: dts: ti: k3-am642-evm: Add M4F remoteproc node
The AM64x SoCs of the TI K3 family have a Cortex M4F core in the MCU
domain. This core can be used by non safety applications as a remote
processor. When used as a remote processor with virtio/rpmessage IPC,
two carveout reserved memory nodes are needed. The first region is used
as a DMA pool for the rproc device, and the second region will furnish
the static carveout regions for the firmware memory.

The current carveout addresses and sizes are defined statically for
each rproc device. The M4F processor does not have an MMU, and as such
requires the exact memory used by the firmware to be set-aside.

Change-Id: Ib096a0d3e5659d696b89efb6710b28bb95909dad
Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Hari Nagalla <hnagalla@ti.com>
(cherry picked from commit e1d2006a5b8ce1b915e577bd2789248bfd0b9c81)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>