TI: dt-bindings: PCI: ti,j721e-pci-host: Add ACSPCIE proxy control property
Add the "ti,syscon-acspcie-proxy-ctrl" device-tree property which is
used to obtain a reference to the ACSPCIE Proxy Control register along
with the details of the PAD IO Buffer output enable bits.

The ACSPCIE Proxy Control register is used to drive the reference clock
for the PCIe Endpoint device via the PAD IO Buffers of the ACSPCIE module.
The ACSPCIE module can be used as an alternative to either an on-board
clock generator or an external clock generator for providing the reference
clock to the PCIe Endpoint device.

Change-Id: I256126b92ef086330f69d462908822bd92cb9977
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
(cherry picked from commit 1b09322960a67a681a2682edd66bf34fe7e8b354)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>