TI: spi: cadence-quadspi: fix error check when finding rxhigh
During PHY calibration the return code of cqspi_find_rx_high() is not
assigned to 'ret'. So the check for failure will take the previous value
of 'ret', which will always be 0 since it was already checked by the
previous step. This means that the calibration procedure will always
think that rxhigh is found even when it is not. Fix this by assigning
the return value of cqspi_find_rx_high() to 'ret'.

Fixes: be58a6d2184f ("spi: cadence-qspi: Tune PHY to allow running at higher frequencies")
Reported-by: Brian Paiva <bpaiva@blackberry.com>
Change-Id: I52940d24b9b59d2b5fffe32af0a7da613ba77e1c
Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Apurva Nandan <a-nandan@ti.com>
Signed-off-by: Santhosh Kumar K <s-k6@ti.com>
(cherry picked from commit 0b5c9cbb6f301a390ffb1e0e6dd86e33187c3cb7)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>