TI: remoteproc: k3-r5: Introduce PM suspend/resume handlers
Introduce suspend/resume handling to r5 remoteproc driver. The driver
registers for pm notifications. And on SUSPEND_PREPARE event a i
RP_MBOX_SUSPEND_SYSTEM mailbox message is sent to the remote core.
a) If the remote core does n't respond to the SUSPEND message, the
driver sets a device on constraint and returns.
b) If the remote core responds with 'RP_MBOX_SUSPEND_ACK', the driver
initiates a stop of the remote core and returns from the suspend
handler.
c) On the other hand, if the remote core responds with
RP_MBOX_SUSPEND_AUTO or RP_MBOX_SUSPEND_CANCEL, the driver simply
returns.

On the Resume path, driver queries DM to determine the state of the
remote core and if it is Off, turns on the remote core.

Change-Id: If1a8baaba2a48bc93d6bf77a055c87aa7a46a506
Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
(cherry picked from commit 1dc201a1afb1fbf1fc5491706cfce40ed6396391)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>